Creates App.jsx
